Added <release:update-version> feature
This is an easier way to update version numbers, and is more portable as
you don't need "sed". Just pass the file to update and a regular
expression matching the version assignment. The first match group will
be replaced with the version. Example:

  <release:update-version path='main.py'>^version = '(.*)'$</release:update-version>

Note: this is sets a "-post" version after the commit, which wasn't
previously possible.